requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
djoser==2.3.1
python-dotenv==1.0.1
sqlparse==0.5.2
tzdata==2024.2
